Separate menus into their own respective classes
