+++ src/com/tti/Main.java
@@ -5,9 +5,11 @@ public class Main {
 	public static void main(String[] args) {
 				
 //		Test.executeAllTests();
+		
 		VehicleInventory.populateHashMap();
+		
 		while (true) {
-		Menu.loadMainMenu();
+			MainMenu.load();
 		}
 	}
